+++ source/logging/StdOutLogger.cpp
@@ -36,7 +36,7 @@ void StdOutLogger::run()
 
 bool StdOutLogger::start(const PlainConfig &config)
 {
-    setLogLevel(config.logConfig.logLevel);
+    setLogLevel(config.logConfig.deviceClientlogLevel);
 
     thread log_thread(&StdOutLogger::run, this);
     log_thread.detach();